LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-construction.any-expected.txt


FAIL Test we can construct a VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test closed VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test we can construct a VideoFrame with a negative timestamp. Can't find variable: OffscreenCanvas
FAIL Test that timestamp is required when constructing VideoFrame from ImageBitmap promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
FAIL Test that timestamp is required when constructing VideoFrame from OffscreenCanvas promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
FAIL Test that timestamp is NOT required when constructing VideoFrame from another VideoFrame promise_test: Unhandled rejection with value: object "NotSupportedError: VideoPixelFormat is not supported"
FAIL Test we can construct an odd-sized VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test constructing w/ unusable image argument throws: HAVE_NOTHING <video>. assert_throws_dom: function "() => {
    let frame = new VideoFrame(video, {timestamp: 10});
  }" threw object "TypeError: Member VideoFrameInit.visibleRect is required and must be an instance of DOMRectInit" that is not a DOMException InvalidStateError: property "code" is equal to undefined, expected 11
FAIL Test constructing w/ unusable image argument throws: emtpy Canvas. Can't find variable: OffscreenCanvas
FAIL Test constructing w/ unusable image argument throws: closed ImageBitmap. Can't find variable: OffscreenCanvas
FAIL Test constructing w/ unusable image argument throws: closed VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test invalid CanvasImageSource constructed VideoFrames VideoPixelFormat is not supported
FAIL Test visibleRect metadata override where source display size = visible size VideoPixelFormat is not supported
FAIL Test visibleRect metadata override where source display width = 2 * visible width (anamorphic) VideoPixelFormat is not supported
FAIL Test visibleRect metadata override where source display size = 2 * visible size for both width and height VideoPixelFormat is not supported
FAIL Test visibleRect + display size metadata override Can't find variable: OffscreenCanvas
FAIL Test display size metadata override Can't find variable: OffscreenCanvas
FAIL Test invalid buffer constructed VideoFrames assert_throws_js: invalid coded size function "() => constructFrame({
                   timestamp: 1234,
                   codedWidth: Math.pow(2, 32) - 1,
                   codedHeight: Math.pow(2, 32) - 1,
                 })" threw object "NotSupportedError: VideoPixelFormat is not supported" ("NotSupportedError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Test Uint8Array(ArrayBuffer) constructed I420 VideoFrame VideoPixelFormat is not supported
FAIL Test ArrayBuffer constructed I420 VideoFrame VideoPixelFormat is not supported
FAIL Test planar constructed I420 VideoFrame with colorSpace VideoPixelFormat is not supported
FAIL Test buffer constructed I420+Alpha VideoFrame VideoPixelFormat is not supported
FAIL Test buffer constructed NV12 VideoFrame assert_equals: plane format expected (string) "NV12" but got (object) null
FAIL Test buffer constructed RGB VideoFrames assert_equals: plane format expected (string) "RGBA" but got (object) null
FAIL Test VideoFrame constructed VideoFrame Can't find variable: OffscreenCanvas
FAIL Test we can construct a VideoFrame from an offscreen canvas. Can't find variable: OffscreenCanvas
FAIL Test I420 VideoFrame with odd visible size VideoPixelFormat is not supported
FAIL Test I420A VideoFrame and alpha={keep,discard} VideoPixelFormat is not supported
FAIL Test RGBA, BGRA VideoFrames with alpha={keep,discard} assert_equals: plane format expected (string) "RGBA" but got (object) null
FAIL Test a VideoFrame constructed from canvas can drop the alpha channel. Can't find variable: OffscreenCanvas
